changing id from string to int

+++ core/repository/mastodon/src/main/java/org/mozilla/social/core/repository/mastodon/model/notifications/NetworkToExternal.kt
@@ -7,47 +7,47 @@ import org.mozilla.social.core.repository.mastodon.model.status.toExternalModel
 fun NetworkNotification.toExternal(): Notification =
     when (this) {
         is NetworkNotification.Mention -> Notification.Mention(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
         )
         is NetworkNotification.NewStatus -> Notification.NewStatus(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
         )
         is NetworkNotification.Repost -> Notification.Repost(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
         )
         is NetworkNotification.Follow -> Notification.Follow(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
         )
         is NetworkNotification.FollowRequest -> Notification.FollowRequest(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
         )
         is NetworkNotification.Favorite -> Notification.Favorite(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
         )
         is NetworkNotification.PollEnded -> Notification.PollEnded(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
         )
         is NetworkNotification.StatusUpdated -> Notification.StatusUpdated(
-            id = id,
+            id = id.toInt(),
             createdAt = createdAt,
             account = account.toExternalModel(),
             status = status.toExternalModel(),
